+++ src/test/java/FeatureFiles/_05_UserStory.feature
@@ -1,0 +1,6 @@
+Feature: Add - Edit - Delete Positions Under Human Resources Setup
+
+  Scenario: Login
+    Given Navigate to Campus
+    When Enter username and password and click login button
+    Then User should login successfully
