--- src/test/java/StepDefinitions/StepDefinitions_10_Steps/User_Story_10_DeleteSteps.java
@@ -1,21 +1,0 @@
-package StepDefinitions.StepDefinitions_10_Steps;
-
-import Pages.DialogContent;
-import Pages.Parent;
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.When;
-
-public class User_Story_10_DeleteSteps {
-
-    DialogContent dc=new DialogContent();
-    Parent parent=new Parent();
-
-    @Given("Deletion is done by clicking the Setup, Parameters, Grade Levels options from the menu categories.")
-    public void deletionIsDoneByClickingTheSetupParametersGradeLevelsOptionsFromTheMenuCategories() {
-    }
-
-    @When("Verify Class Levels Have Been Deleted")
-    public void verifyClassLevelsHaveBeenDeleted() {
-    }
-
-}
