--- src/test/java/StepDefinitions/StepDefinitions_10_Steps/User_Story_10_EditSteps.java
@@ -1,24 +1,0 @@
-package StepDefinitions.StepDefinitions_10_Steps;
-
-import Pages.DialogContent;
-import Pages.FormContent;
-import Pages.Parent;
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.When;
-
-public class User_Story_10_EditSteps {
-
-    DialogContent dc=new DialogContent();
-    FormContent fm=new FormContent();
-    Parent parent=new Parent();
-
-    @Given("by clicking the Setup, Parameters, Grade Levels options.")
-    public void byClickingTheSetupParametersGradeLevelsOptions() {
-    }
-
-    @When("confirm grade levels")
-    public void confirmGradeLevels() {
-    }
-
-
-}
