--- src/test/java/StepDefinitions/StepDefinitions_10_Steps/User_Story_10_addSteps.java
@@ -1,22 +1,0 @@
-package StepDefinitions.StepDefinitions_10_Steps;
-
-import Pages.DialogContent;
-import Pages.FormContent;
-import Pages.LeftNav;
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.When;
-
-public class User_Story_10_addSteps {
-    DialogContent dc=new DialogContent();
-    LeftNav ln=new LeftNav();
-    FormContent fm=new FormContent();
-
-    @Given("Click Setup, Parameters, Class Levels in the menu category.")
-    public void clickSetupParametersClassLevelsInTheMenuCategory() {
-    }
-
-    @When("that grade levels have been added")
-    public void thatGradeLevelsHaveBeenAdded() {
-    }
-
-}
