--- src/test/java/StepDefinitions/StepDefinitions_10_Steps/case10.java
@@ -1,47 +1,0 @@
-package StepDefinitions.StepDefinitions_10_Steps;
-
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
-import io.cucumber.java.en.When;
-
-public class case10 {
-    @Given("Go to Given Site")
-    public void goToGivenSite() {
-    }
-
-    @When("Enter the username and password and click the login button")
-    public void enterTheUsernameAndPasswordAndClickTheLoginButton() {
-    }
-
-    @Given("Click Setup, Parameters, Class Levels in the menu category.")
-    public void clickSetupParametersClassLevelsInTheMenuCategory() {
-    }
-
-    @When("that grade levels have been added")
-    public void thatGradeLevelsHaveBeenAdded() {
-    }
-
-    @Then("confirm successfully")
-    public void confirmSuccessfully() {
-    }
-
-    @Given("by clicking the Setup, Parameters, Grade Levels options.")
-    public void byClickingTheSetupParametersGradeLevelsOptions() {
-    }
-
-    @When("confirm grade levels")
-    public void confirmGradeLevels() {
-    }
-
-    @Then("accurately and successfully")
-    public void accuratelyAndSuccessfully() {
-    }
-
-    @Given("Deletion is done by clicking the Setup, Parameters, Grade Levels options from the menu categories.")
-    public void deletionIsDoneByClickingTheSetupParametersGradeLevelsOptionsFromTheMenuCategories() {
-    }
-
-    @When("Verify Class Levels Have Been Deleted")
-    public void verifyClassLevelsHaveBeenDeleted() {
-    }
-}
