+++ src/test/java/StepDefinitions/_01_UserStorySteps/User_Story_1_EditSteps.java
@@ -1,23 +1,23 @@
-package StepDefinitions.User_Story_6_Steps;
+package StepDefinitions._01_UserStorySteps;
 
 import Pages.DialogContent;
 import Pages.FormContent;
 import Pages.Parent;
 import io.cucumber.java.en.When;
 
-public class User_Story_6_EditSteps {
+public class User_Story_1_EditSteps {
     DialogContent dc=new DialogContent();
     FormContent fm=new FormContent();
     Parent parent=new Parent();
-    @When("Edit to Subject Categories")
-    public void editToSubjectCategories() {
-        dc.findAndSend("NAMEINPUT","sude");
-        dc.findAndClick("searchbutton");
-        dc.findAndClick("EDİTBUTTON");
-        fm.findAndSend("nameInput","enesff");
-        fm.findAndSend("code","009988");
-        dc.findAndClick("saveButton");
+    @When("Edit to position category")
+    public void editToPositionCategory() {
 
+        dc.findAndSend("NAMEINPUT","Ahmet22");
+        dc.findAndClick("searchbutton");
+        parent.waitUntilLoading();
+        dc.findAndClick("EDİTBUTTON");
+        fm.findAndSend("nameInput","Tunaş");
+        dc.findAndClick("saveButton");
 
     }
 }
