+++ src/test/java/StepDefinitions/_06_UserStorySteps/User_Story_6_DeleteSteps.java
@@ -1,4 +1,4 @@
-package StepDefinitions.User_Story_6_Steps;
+package StepDefinitions._06_UserStorySteps;
 
 import Pages.DialogContent;
 import Pages.Parent;
