+++ src/test/java/StepDefinitions/_06_UserStorySteps/User_Story_6_EditSteps.java
@@ -1,23 +1,23 @@
-package StepDefinations.Story1;
+package StepDefinitions._06_UserStorySteps;
 
 import Pages.DialogContent;
 import Pages.FormContent;
 import Pages.Parent;
 import io.cucumber.java.en.When;
 
-public class User_Story_1_EditSteps {
+public class User_Story_6_EditSteps {
     DialogContent dc=new DialogContent();
     FormContent fm=new FormContent();
     Parent parent=new Parent();
-    @When("Edit to position category")
-    public void editToPositionCategory() {
-
-        dc.findAndSend("NAMEINPUT","Ahmet22");
+    @When("Edit to Subject Categories")
+    public void editToSubjectCategories() {
+        dc.findAndSend("NAMEINPUT","sude");
         dc.findAndClick("searchbutton");
-        parent.waitUntilLoading();
         dc.findAndClick("EDİTBUTTON");
-        fm.findAndSend("nameInput","Tunaş");
+        fm.findAndSend("nameInput","enesff");
+        fm.findAndSend("code","009988");
         dc.findAndClick("saveButton");
+
 
     }
 }
