+++ src/test/java/StepDefinitions/_11_UserStorySteps/User_Story_11_DeleteSteps.java
@@ -1,15 +1,17 @@
-package StepDefinitions.Story4;
+package StepDefinitions._11_UserStorySteps;
 
 import Pages.DialogContent;
 import Pages.Parent;
 import io.cucumber.java.en.When;
 
-public class User_Story_4_DeleteSteps {
+public class User_Story_11_DeleteSteps {
     DialogContent dc=new DialogContent();
     Parent parent=new Parent();
-    @When("Delete to Fields")
-    public void deleteToFields() {
-        dc.findAndSend("NAMEINPUT","Ramazan");
+
+    @When("Delete to Discounts")
+    public void deleteToDiscounts() {
+        parent.waitEditButtonVerify();
+        dc.findAndSend("descriptionInput","Tuna");
         dc.findAndClick("searchbutton");
         parent.waitUntilLoading();
         dc.findAndClick("DELETEBUTTON");
